75 - Refactor to get rid of sonar warnings
